Split out all the separate codemirror editors
