+++ studio/src/components/CodeMirrorEditor/index.ts
@@ -1,7 +1,4 @@
-export {
-  CodeMirrorTypescriptEditor,
-  CodeMirrorJsonEditor,
-  CodeMirrorSqlEditor,
-} from "./CodeMirrorEditor";
-
+export { CodeMirrorSqlEditor } from "./CodeMirrorSqlEditor";
+export { CodeMirrorTypescriptEditor } from "./CodeMirrorTypescriptEditor";
+export { CodeMirrorJsonEditor } from "./CodeMirrorJsonEditor";
 export { CodeMirrorInput } from "./CodeMirrorInput";
