+++ studio/src/components/CodeMirrorEditor/themes.ts
@@ -1,0 +1,9 @@
+import { EditorView } from "@uiw/react-codemirror";
+
+export const customTheme = EditorView.theme({
+  "&": {
+    fontSize: "14px",
+    // HACK
+    background: "transparent !important",
+  },
+});
