+++ studio/src/components/CodeMirrorEditor/types.ts
@@ -1,0 +1,9 @@
+export type CodeMirrorEditorProps = {
+  height?: string;
+  minHeight?: string;
+  maxHeight?: string;
+  theme?: string;
+  readOnly?: boolean;
+  value?: string;
+  onChange: (value?: string) => void;
+};
